+++ ubo_app/services/010-ethernet/ubo_handle.py
@@ -6,6 +6,5 @@ from ubo_app.load_services import register_service
 register_service(
     service_id='ethernet',
     label='Ethernet',
+    init=init_service,
 )
-
-init_service()