+++ ubo_app/services/010-ip/ubo_handle.py
@@ -8,6 +8,5 @@ register_service(
     service_id='ip',
     label='IP',
     reducer=reducer,
+    init=init_service,
 )
-
-init_service()