+++ ubo_app/services/010-wifi/pages/create_wireless_connection.py
@@ -31,8 +31,10 @@ from ubo_app.utils.async_ import create_task
 
 # Regular expression pattern
 # WIFI:S:<SSID>;T:<WEP|WPA|blank>;P:<PASSWORD>;H:<true|false|blank>;;
-barcode_pattern = r"""WIFI:S:(?P<SSID>[^;]*);(?:T:(?P<Type>(?i:WEP|WPA|WPA2|nopass));)\
-?(?:P:(?P<Password>[^;]*);)?(?:H:(?P<Hidden>(?i:true|false));)?"""
+barcode_pattern = (
+    r'WIFI:S:(?P<SSID>[^;]*);(?:T:(?P<Type>(?i:WEP|WPA|WPA2|nopass));)'
+    r'?(?:P:(?P<Password>[^;]*);)?(?:H:(?P<Hidden>(?i:true|false));)?'
+)
 
 
 class CreateWirelessConnectionPage(PageWidget):