+++ ubo_app/services/010-wifi/ubo_handle.py
@@ -8,6 +8,5 @@ register_service(
     service_id='wifi',
     label='WiFi',
     reducer=reducer,
+    init=init_service,
 )
-
-init_service()