+++ ubo_app/services/040-camera/ubo_handle.py
@@ -8,6 +8,5 @@ register_service(
     service_id='camera',
     label='Camera',
     reducer=reducer,
+    init=init_service,
 )
-
-init_service()