+++ ubo_app/services/040-sensors/ubo_handle.py
@@ -8,6 +8,5 @@ register_service(
     service_id='sensors',
     label='Sensors',
     reducer=reducer,
+    init=init_service,
 )
-
-init_service()