+++ ubo_app/services/080-docker/ubo_handle.py
@@ -8,6 +8,5 @@ register_service(
     service_id='docker',
     label='Docker',
     reducer=reducer,
+    init=init_service,
 )
-
-init_service()